Add books components which contain book item and booklist
